Add root task (mix hex), replaces hex.info
